requirements.txt
streamlit
youtube-transcript-api
langchain
langchain-community
faiss-cpu
sentence-transformers
groq
python-dotenv
pickle5
pypickle